+++ SOCEngine/SOCEngine/Math/Matrix.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Common.h"
 #include "Rect.h"
 
 namespace Math
@@ -41,8 +42,8 @@ namespace Math
 		static Matrix OrthoLH(float w, float h, float zn, float zf);
 		static Matrix RotateUsingAxis(const Vector3& v, float angle);
 		static Matrix ComputeViewMatrix(const Matrix &worldMatrix);
-		static Matrix ComputeViewportMatrix(const Rect<float>& rect);
-		static Matrix ComputeInvViewportMatrix(const Rect<float>& rect);
+		static Matrix ComputeViewportMatrix(const Rect<uint>& rect);
+		static Matrix ComputeInvViewportMatrix(const Rect<uint>& rect);
 		static Matrix MakeRotationMatrix(const Vector3& right, const Vector3& up, const Vector3& forward);
 		static Matrix LookAtDir(const Vector3 & targetDir, const Vector3* upVec = nullptr);
